Refactor rock-blockwise-gemm-to-threadwise pass
