+++ mlir/include/mlir/Dialect/Rock/IR/RockOps.td
@@ -807,8 +807,8 @@ def Rock_IndexDiffUpdateOp :
 defvar SupportedMemoryElems = [F32, F16, BF16, I8, I32, F8E5M2FNUZ, F8E4M3FNUZ];
 defvar NativeMemoryOpTypes = [F32, F16, BF16, I8, I32,
                            F8E5M2FNUZ, F8E4M3FNUZ,
-                           VectorOfLengthAndType<[2, 4], [F32, I32]>,
-                           VectorOfLengthAndType<[2, 4, 8], [F16, BF16]>,
+                           VectorOfLengthAndType<[2, 4, 8], [F32, I32]>,
+                           VectorOfLengthAndType<[2, 4, 8, 16], [F16, BF16]>,
                            VectorOfLengthAndType<[2, 4, 8, 16],
                              [I8, F8E5M2FNUZ, F8E4M3FNUZ]>];
 
@@ -935,9 +935,8 @@ def Rock_GlobalStoreOp :
 def Rock_ThreadwiseReadIntoOp :
     Rock_Op<"threadwise_read_into">,
     AllElementTypesMatch<["source", "dest"]>,
-    Arguments<(ins Arg<MemRefOf<SupportedMemoryElems>, "source view", [MemRead]>:$source,
-      Arg<MemRefRankOf<SupportedMemoryElems, [1]>,
-        "destination registers", [MemWrite]>:$dest,
+    Arguments<(ins Arg<MemRefOf<NativeMemoryOpTypes>, "source view", [MemRead]>:$source,
+      Arg<MemRefOf<NativeMemoryOpTypes>, "destination registers", [MemWrite]>:$dest,
       TransformMapArrayAttr:$extraViews,
       Variadic<Index>:$extraIndices,
       UnitAttr:$forceUnroll,
